+++ src/main/java/tornadofx/Nodes.kt
@@ -2,6 +2,7 @@ package tornadofx
 
 import com.sun.javafx.scene.control.skin.TableColumnHeader
 import javafx.application.Platform
+import javafx.beans.value.ObservableValue
 import javafx.collections.FXCollections.observableArrayList
 import javafx.collections.ObservableList
 import javafx.event.EventTarget
@@ -794,4 +795,20 @@ fun EventTarget.removeFromParent() {
     } else if (this is Node) {
         parent?.getChildList()?.remove(this)
     }
+}
+
+/**
+ * Listen for changes to an observable value and replace all content in this Node with the
+ * new content created by the onChangeBuilder. The builder operates on the node and receives
+ * the new value of the observable as it's only parameter.
+ *
+ * The onChangeBuilder is run immediately with the current value of the property.
+ */
+fun <S : EventTarget, T>S.dynamicContent(property: ObservableValue<T>, onChangeBuilder: S.(T?) -> Unit) {
+    val onChange : (T?) -> Unit = {
+        getChildList()?.clear()
+        onChangeBuilder(this@dynamicContent, it)
+    }
+    property.onChange(onChange)
+    onChange(property.value)
 }